Add scaffolding for the NPM 7 project (#5)
# Description

This code update includes:
- repository documentation
- github templates and workflows 
- codeowners file
- publishing configuration files
- linting configurations

This change uses the same format as the https://github.com/microsoft/fast-tooling repository, requiring NPM 7+ and NodeJs 16+.

## Issues

Closes #4 

## Reviewers

Note that this initial PR will fail as there is no workspaces to test against. This should be fine and expected as the next pull request will add a package.