Added configuration command and cleaned up init (#33)
# Pull Request

## 📖 Description

<!--- Provide some background and a description of your work. -->
This change adds the `config` command which creates a config file with a component path. It also cleans up some of the `init` logic where template files provided in the command line were assumed to be local packages.

### 🎫 Issues

<!---
List and link relevant issues here using the keyword "closes"
if this PR will close an issue, eg. closes #411
-->

Closes #27 

## ✅ Checklist

### General

<!--- Review the list and put an x in the boxes that apply. -->

- [x] I have added tests for my changes.
- [x] I have tested my changes.
- [x] I have updated the project documentation to reflect my changes.

## ⏭ Next Steps

<!---
If there is relevant follow-up work to this PR, please list any existing issues or provide brief descriptions of what you would like to do next.
-->
- #28