Updated the specification around adding foundation components (#37)
# Pull Request

## 📖 Description

<!--- Provide some background and a description of your work. -->
This update to the specification clarifies adding a "foundation" component with a specific `add-foundation-component` command.

## 👩‍💻 Reviewer Notes

<!---
Provide some notes for reviewers to help them provide targeted feedback and testing.
-->
Read through and make sure this command makes sense given the `add-component` command. Since the `add-component` command deals with adding via templates, the `add-foundation-component` is intended to explicit to adding foundation based component. This distinction was thought to be necessary as this adds the `@microsoft/fast-foundation` dependency and begins to use the preconfigured design tokens, so if you are not intending to use those systems, you would simply use the `add-component` command for your custom components.

## ✅ Checklist

### General

<!--- Review the list and put an x in the boxes that apply. -->

- [ ] I have added tests for my changes.
- [ ] I have tested my changes.
- [x] I have updated the project documentation to reflect my changes.

## ⏭ Next Steps

<!---
If there is relevant follow-up work to this PR, please list any existing issues or provide brief descriptions of what you would like to do next.
-->
- #31