Add the add-foundation-component prompt (#38)
# Pull Request

## 📖 Description

<!--- Provide some background and a description of your work. -->
This update does the following:
- Adds the `add-foundation-component` command
- Adds the "blank" template to the list of foundation components
- Updates the documentation to normalize it
- Adds the documentation for the `add-foundation-component` command

### 🎫 Issues

<!---
List and link relevant issues here using the keyword "closes"
if this PR will close an issue, eg. closes #411
-->
Partially takes care of #31

## 👩‍💻 Reviewer Notes

<!---
Provide some notes for reviewers to help them provide targeted feedback and testing.
-->
In order to not have a massive PR down the line, foundation components will be added individually as PRs. This PR only takes care of the command line option and moves the tests around to better facilitate this update.

## ✅ Checklist

### General

<!--- Review the list and put an x in the boxes that apply. -->

- [x] I have added tests for my changes.
- [x] I have tested my changes.
- [x] I have updated the project documentation to reflect my changes.

## ⏭ Next Steps

<!---
If there is relevant follow-up work to this PR, please list any existing issues or provide brief descriptions of what you would like to do next.
-->
- Add each foundation component