fix(docs): add-a-component command argument typo fix (#93)
I updated the syntax of the sentence and fixed a typo assuming that what's meant is **pre-configured** not **pre-congured**.

# Pull Request

## 📖 Description

Docs update to fix typo in add-component page when explaining the arguments that should be supplied to the CLI command during component creation. 
<!--- Provide some background and a description of your work. -->

### 🎫 Issues
Not applicable
<!---
List and link relevant issues here using the keyword "closes"
if this PR will close an issue, eg. closes #411
-->

## 👩‍💻 Reviewer Notes
Not applicable
<!---
Provide some notes for reviewers to help them provide targeted feedback and testing.
-->

## ✅ Checklist

### General

<!--- Review the list and put an x in the boxes that apply. -->

- [*] I have updated the project documentation to reflect my changes.

## ⏭ Next Steps
Not applicable
<!---
If there is relevant follow-up work to this PR, please list any existing issues or provide brief descriptions of what you would like to do next.
-->

Co-authored-by: Jane Chu <7559015+janechu@users.noreply.github.com>