Added migrate package for FAST CLI (#104)
# Pull Request

## 📖 Description

<!--- Provide some background and a description of your work. -->
This change adds:
- The migrate ESLint package which contains fixes for deprecated code from previous versions
- A rule for migrating component class files

### 🎫 Issues

<!---
List and link relevant issues here using the keyword "closes"
if this PR will close an issue, eg. closes #411
-->
Begins work on #41 

## 👩‍💻 Reviewer Notes

<!---
Provide some notes for reviewers to help them provide targeted feedback and testing.
-->
This new package will be leveraged by the FAST CLI, if the user chooses they can either attempt to fix, which will fire the fix command against their code and update it, or they can simply run it and get an output of all the issues detected.

The package will consume the `fast.config.json` file and therefore should know the locations of components and various other files needed for the migration to work.

You will see a mention of a "version matrix" in the `README.md` file. A single version will be stored in the `fast.config.json` and based on this version we will determine what versions of `@microsoft/fast-element` and `@microsoft/fast-foundation` the project should be on. This allows us to understand our upgrade migration path for both dependencies.

## ✅ Checklist

### General

<!--- Review the list and put an x in the boxes that apply. -->

- [x] I have added tests for my changes.
- [x] I have tested my changes.
- [x] I have updated the project documentation to reflect my changes.

## ⏭ Next Steps

<!---
If there is relevant follow-up work to this PR, please list any existing issues or provide brief descriptions of what you would like to do next.
-->
- Continue adding migrate rules